Update sketch.js for lake and land
I started with the previous sketches of the lake and land for level 1. To add levels 2-4, and to keep them compatible with the group code, I created two offscreen canvases: pg still stores the original level 1 image, and pg2 is rebuilt on each level 2-4 button click to show the new strokes. I replaced the old currentExpression string with a number, activeLevel. When the button is pressed, the activeLevel is updated, and if it is a higher level, pg2 is rebuilt, and the draw() loop displays the desired canvas.
The remaining problem is that the lake and land outlines still fail after resizing; I will focus on fixing the scaling problem of lake and land next.